Extract API classes to a separate module
