+++ script-generator/web/api/src/commonMain/kotlin/io/github/typesafegithub/workflows/scriptgenerator/rest/api/YamlToKotlinResponse.kt
@@ -1,0 +1,8 @@
+package io.github.typesafegithub.workflows.scriptgenerator.rest.api
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class YamlToKotlinResponse(
+    val kotlinScript: String,
+)
