+++ script-generator/web/ui/build.gradle.kts
@@ -6,6 +6,8 @@ plugins {
 }
 
 dependencies {
+    implementation(projects.scriptGenerator.web.api)
+
     implementation(enforcedPlatform("org.jetbrains.kotlin-wrappers:kotlin-wrappers-bom:1.0.0-pre.542"))
 
     implementation("org.jetbrains.kotlin-wrappers:kotlin-react")
